requirements-dev.txt
# Development dependencies for CI/CD
black==24.10.0
ruff==0.8.4
pytest==8.3.4
pytest-django==4.9.0
coverage==7.6.9
bandit==1.7.10
pip-audit==2.7.3
